+++ bee-common/bee-common/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Packable` implementation for `Vec<T: Packable>`
 - Add `filters` option to the logger.
+- Added event bus
 
 ### Changed